Read the FFmpeg command duration as a decimal
It was read with atoi(), so ffprobe's "12.512000" became 12 whole
seconds. Files that report a frame count of their own never used it, but
those that do not -- MXF among them -- work the duration out from it, so
their duration, start time and speed all came out wrong. The audio half
already used atof().

+++ lib/tlRender/IO/FFmpegCmdRead.cpp
@@ -363,7 +363,11 @@ namespace tl
                     if (auto j = i.value().find("duration");
                         j != i.value().end())
                     {
-                        duration = atoi(j->get<std::string>().c_str());
+                        // Not atoi(): ffprobe gives this as a decimal, and
+                        // truncating it to whole seconds is what made the
+                        // duration wrong for the files that have no frame
+                        // count of their own to use instead.
+                        duration = atof(j->get<std::string>().c_str());
                     }
 
                     // Get the metadata tags.
